authoptions in separate file
