+++ app/components/buttons/actions.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import authOptions from '@/app/api/auth/authoptions';
 import prisma from '@/db';
 import { getServerSession } from 'next-auth';
 import { revalidatePath } from 'next/cache';
