+++ app/event_series/page.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../api/auth/[...nextauth]/route';
+import authOptions from '../api/auth/authoptions';
 import {
   getFilteredEventSeries,
   getTotalPages,
